requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
blinker==1.5
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==37.0.4
distlib==0.3.6
Django==4.1.3
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
greenletio==0.9.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
mysql==0.0.3
mysqlclient==2.1.1
outcome==1.2.0
packaging==21.3
passlib==1.7.4
platformdirs==2.5.4
protobuf==4.21.9
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.21.1
requests==2.28.2
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2
SQLAlchemy==1.4.44
sqlparse==0.4.3
stripe==5.0.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.14
urwid==2.1.2
uvicorn==0.20.0
virtualenv==20.16.7
webdriver-manager==3.8.1
Werkzeug==2.2.2
wsproto==1.1.0
environs==9.5.0
pdfkit==1.0.0
boto3==1.26.107
gunicorn==20.1.0
wheel==20.1.0
uwsgi==2.0.21
